Adopter Profile
Create Adopter_profiles_controller and REST actions for new, show, edit, create and associated views. Also add logic to home page for the Profile link to not show it if not an adopter user, and to go to new profile form if one does not exist, or show existing profile if one exists.