Nested attributes
Add nested attributes to User model for staff_account and adopter_account and after testing it does create either a staff or adopter account upon new registration. This required a #new action in the registrations controller to create new instances of each account model, and addition of field_for in the registrations>new view. More research needed into best practice for creating models this way as I could only get it to work by including the user_id field as a nested attribute, and testing shows despite playing with this number the account appropriately records the corred user_id of the new registrant. Needs logic to select which nested form to include based on user account selection. Also migration to add default organization_id to staff_accounts of 1, and a null false.