Adoptoion application reviews
Add instance variable to #index action to allow for one dog object to be retrieved, or many in an active record relation, and add logic to the partial view to render the appropriate HTML based on whether there is one object or a relation. This allows a user to filter the applications page for a particular dog.