Adoption
Instead of nesting adoption inside of the edit_adopter_application, decided to make a separate form/button for the action to create a new adoption. Add Adoptions controller with Create action and strong params. Add patch route pointing to Adoptions#Create. Add button to adopt in the adopter_application view so it shows on each application row