Adopter Applications
Remove #delete from adopter_applications controller as I decided to not allow deletion of these records. Add column profile_show  to Adopter_applications table that will be used in view logic to determine whether or not to show an application. Add update method to controller and a link to withdraw from an application in the profile _applications view so a user can click withdraw and set the status to withdrawn. Tested and works.