Org dogs controller and index
Add new instance variables to the index action that create three collections: dogs that are adopted, dogs that need adoption, and a single dog. Add conditional logic to index view that sets a collection variable to the appropriate controller instance variable depending on the selection made in the two forms to filter on adoption status and name.