Appending Images
Rails server output had a deprecation notice stating that config.active_storage.replace_on_assign_to_many would be deprecated in rails 7.1 and that the app should work with that config set to true. I had it set to false to allow appending of images. Instead, set it to true and added a model method using the attach method to append attachables to existing attachments.