File upload bug fix
File upload bug occurred when an upload failed validation and was therefore not saved to database, but delete button was calling an id that did not exist on said attachment. Add logic to view to only display image and delete link if the image had a created_at attribute i.e., was saved to the database.