Devise customization
Fix flash messages on failed login or sign up by setting data turbo to false on the submit buttons on each form. This is a workaround solution that is a compromise between speed/complexity and functionality. It was quick and easy to do, and will be easy to remove in future, but stops turbo stream and results in a slower user experience. It is a reasonable solution until I find a more official solution. Add first and last name fields to the edit user settings form.