Fix failing test by leveraging Capybara's waiting mechanism
It seems this test began failing consistently after
updating to the latest Rails patch version.

It appears that Capybara may have been clicking the image too quickly,
perhaps not giving the associated Javascript enough time to load.

Passing in a block when finding the element allows
it to retry if necessary, then pass.

A small benefit of this change is that it does not have to find
the image element again.