Add birth_date to Pet and compute age dynamically
Ruby (https://bugs.ruby-lang.org/issues/14877) nor ActiveSupport have a
way to compute the age given a birthday and most answers online are
concerned with human ages which tend to only be in years. Since it seems
like we want to include months, especially if the pet is less than 1
year old, I added a custom function for computing and display that. It
isn't technically accurate because we compute it off the duration from
today and the birthdate, but should be acceptable because birthdates in
the system will typically be approximate anyway.