Add birth_date to Pet and compute age dynamically (#121)
Closes https://github.com/rubyforgood/pet-rescue/issues/90

Currently age is a static form field, so if a month passes in real time, the pet's age stays the same.

Ruby (https://bugs.ruby-lang.org/issues/14877) nor ActiveSupport have a
way to compute the age given a birthday and most answers online are
concerned with human ages which tend to only be in years. ~~Since it seems
like we want to include months, especially if the pet is less than 1
year old, I added a custom function for computing and display that. It
isn't technically accurate because we compute it off the duration from
today and the birthdate, but should be acceptable because birthdates in
the system will typically be approximate anyway.~~

Switched this to use https://api.rubyonrails.org/classes/ActionView/Helpers/DateHelper.html#method-i-time_ago_in_words, which is still approximate but should serve our purposes. The only difference is that instead of always saying "exact" years and months, it will sometimes use "about", "almost", "over".