Fixes: #754
We wanted to make a change to the pet overview tab when viewing a specific pet. Previous to the change, half of the columns in the overview table were being hidden if the pet had any matches at all. This change makes it so that if there are no matches OR the match type is :foster, the full overview table is still viewable by the staff. The only time part of the overview should be hidden is if a match is made that is the regular :match type, AKA an adopter. I added a check in the '_overview.html.erb' partial for the :foster type match, and added a new test file, 'pet_overview_test.rb' to make sure this functionality was working. I also QA tested manually via the console, and it seems to be working.