sketch out database structure for chat
I'm sure I'll end up rebasing this to add more fields later ...

The base model is a Chat, which is a record of a conversation on a
particular day.  Each chat has multiple Participants; at least one of
these is the staff member of an organization initiating the chat, and
others may be either the Fosterer or potential Adopter.

Each chat consists of multiple messages, in sequence, and can receive a
"summary" describing what the chat was about